First stab at Profiler documentation
Note that the images are only placeholders for now...will refine them a
bit further. Also, the purpose of this documentation is to explain the
tool, not CSS optimisation ins and outs, or how/when browser operations
are actually triggered. For this reason, it's slightly sparse on the WHY
of the result.